Remove tempfile lock before continuing analysis
This fixes #17 that I also ran into. This is just a copied patch from the original repo where it was also added by the original author.

Without this line, ghidra fails to open the project to analyze anything and just makes a mess in the directory.

With this line, it analyzes properly and makes no mess.

Thanks for the excellent videos!